payment_api/requirements.txt
sanic
sqlalchemy[asyncio]
asyncpg
alembic
python-jose[cryptography]
python-dotenv
pydantic>=1.9.0